optimized fetching
After realizing that the hospital data is static and that im a fucking dumbass for not realizing it, I have made it so instead of fetching.js getting the hospital data from the API it gets it from a local json file. This will probably change once we have our own server set up later but for now this *should* run much faster.